Split different APIs into several mixins for clarity / easier maintenance
